Constants and convenience functions

SNMRForward exports values of a number of useful physical constants related to the physics of the SNMR technique.

SNMRForward.μ0 — Constant

magnetic vacuum permeability (Newton/(Ampere^2))

SNMRForward.γh — Constant

gyromagnetic ratio of the hydrogen nucleus (bound in water) (radians/(second Tesla))

SNMRForward.ħ — Constant

reduced Planck's constant (Joule second)

SNMRForward.kb — Constant

boltzmann constant (Joule/Kelvin)

SNMRForward.nh2o — Constant

number density of water (m^(-3))

SNMRForward.mag_factor — Function

m0 = mag_factor(temp)

Returns the ratio between the magnetisation of completely saturated water (100% water content) and the ambient magnetic field at thermal equilibrium.

Parameters:

temp: temperature in Kelvin.

Returns:

m0: magnetisation ratio, in Amperes per (metre Tesla).